refactor: seperate WorkTime and AvailableTime Operations
